Code cleanup. 1. Move DxilEmitMetadata out of DxilGenerationPass.cpp 2. Move helper function out of HLModule.
