+++ include/dxc/HLSL/DxilUtil.h
@@ -1,0 +1,34 @@
+///////////////////////////////////////////////////////////////////////////////
+//                                                                           //
+// DxilUtil.h                                                                //
+// Copyright (C) Microsoft Corporation. All rights reserved.                 //
+// This file is distributed under the University of Illinois Open Source     //
+// License. See LICENSE.TXT for details.                                     //
+//                                                                           //
+// Dxil helper functions.                                                    //
+//                                                                           //
+///////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+namespace llvm {
+class Type;
+class GlobalVariable;
+}
+
+namespace hlsl {
+
+class DxilFieldAnnotation;
+class DxilTypeSystem;
+
+namespace dxilutil {
+  unsigned
+  GetLegacyCBufferFieldElementSize(DxilFieldAnnotation &fieldAnnotation,
+                                   llvm::Type *Ty, DxilTypeSystem &typeSys);
+  llvm::Type *GetArrayEltTy(llvm::Type *Ty);
+
+  bool IsStaticGlobal(llvm::GlobalVariable *GV);
+  bool IsSharedMemoryGlobal(llvm::GlobalVariable *GV);
+}
+
+}
