+++ lib/HLSL/CMakeLists.txt
@@ -20,6 +20,7 @@ add_llvm_library(LLVMHLSL
   DxilModule.cpp
   DxilOperations.cpp
   DxilOutputColorBecomesConstant.cpp
+  DxilPreparePasses.cpp
   DxilRemoveDiscards.cpp
   DxilReduceMSAAToSingleSample.cpp
   DxilPreserveAllOutputs.cpp
@@ -32,6 +33,7 @@ add_llvm_library(LLVMHLSL
   DxilSignature.cpp
   DxilSignatureElement.cpp
   DxilTypeSystem.cpp
+  DxilUtil.cpp
   DxilValidation.cpp
   DxcOptimizer.cpp
   HLMatrixLowerPass.cpp
