+++ lib/HLSL/HLModule.cpp
@@ -791,65 +791,6 @@ bool HLModule::IsHLSLObjectType(llvm::Type *Ty) {
   return false;
 }
 
-Type *HLModule::GetArrayEltTy(Type *Ty) {
-  if (isa<PointerType>(Ty))
-    Ty = Ty->getPointerElementType();
-  while (isa<ArrayType>(Ty)) {
-    Ty = Ty->getArrayElementType();
-  }
-  return Ty;
-}
-
-unsigned
-HLModule::GetLegacyCBufferFieldElementSize(DxilFieldAnnotation &fieldAnnotation,
-                                           llvm::Type *Ty,
-                                           DxilTypeSystem &typeSys) {
-  while (isa<ArrayType>(Ty)) {
-    Ty = Ty->getArrayElementType();
-  }
-
-  // Bytes.
-  unsigned compSize = fieldAnnotation.GetCompType().Is64Bit()?8:4;
-  unsigned fieldSize = compSize;
-  if (Ty->isVectorTy()) {
-    fieldSize *= Ty->getVectorNumElements();
-  } else if (StructType *ST = dyn_cast<StructType>(Ty)) {
-    DxilStructAnnotation *EltAnnotation = typeSys.GetStructAnnotation(ST);
-    if (EltAnnotation) {
-      fieldSize = EltAnnotation->GetCBufferSize();
-    } else {
-      // Calculate size when don't have annotation.
-      if (fieldAnnotation.HasMatrixAnnotation()) {
-        const DxilMatrixAnnotation &matAnnotation =
-            fieldAnnotation.GetMatrixAnnotation();
-        unsigned rows = matAnnotation.Rows;
-        unsigned cols = matAnnotation.Cols;
-        if (matAnnotation.Orientation == MatrixOrientation::ColumnMajor) {
-          rows = cols;
-          cols = matAnnotation.Rows;
-        } else if (matAnnotation.Orientation != MatrixOrientation::RowMajor) {
-          // Invalid matrix orientation.
-          fieldSize = 0;
-        }
-        fieldSize = (rows - 1) * 16 + cols * 4;
-      } else {
-        // Cannot find struct annotation.
-        fieldSize = 0;
-      }
-    }
-  }
-  return fieldSize;
-}
-
-bool HLModule::IsStaticGlobal(GlobalVariable *GV) {
-  return GV->getLinkage() == GlobalValue::LinkageTypes::InternalLinkage &&
-         GV->getType()->getPointerAddressSpace() == DXIL::kDefaultAddrSpace;
-}
-
-bool HLModule::IsSharedMemoryGlobal(llvm::GlobalVariable *GV) {
-  return GV->getType()->getPointerAddressSpace() == DXIL::kTGSMAddrSpace;
-}
-
 void HLModule::GetParameterRowsAndCols(Type *Ty, unsigned &rows, unsigned &cols,
                                        DxilParameterAnnotation &paramAnnotation) {
   if (Ty->isPointerTy())
